backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
dnspython
boto3
python-dotenv
langdetect
